+++ src/main/kotlin/ru/mipt/npm/gradle/KScienceProjectPlugin.kt
@@ -8,29 +8,52 @@ import org.gradle.kotlin.dsl.*
 import org.jetbrains.changelog.ChangelogPlugin
 import org.jetbrains.dokka.gradle.DokkaPlugin
 import org.jetbrains.dokka.gradle.DokkaTask
+import ru.mipt.npm.gradle.internal.addGithubPublishing
+import ru.mipt.npm.gradle.internal.addSonatypePublishing
+import ru.mipt.npm.gradle.internal.addSpacePublishing
+import ru.mipt.npm.gradle.internal.setupPublication
 
 @Suppress("unused")
 class KSciencePublishingExtension(val project: Project) {
-    var vcs: String? by project.extra
+    private var initializedFlag = false
 
-    // github publishing
-    var githubOrg: String? by project.extra
-    var githubProject: String? by project.extra
+    fun setup(vcsUrl: String){
+        project.setupPublication(vcsUrl)
+        initializedFlag = true
+    }
 
-    // Space publishing
-    var spaceRepo: String? by project.extra
-    var spaceUser: String? by project.extra
-    var spaceToken: String? by project.extra
+    /**
+     * github publishing
+     */
+    fun github(githubProject: String, githubOrg: String = "mipt-npm") {
+        //automatically initialize vcs using github
+        if(!initializedFlag){
+            setup("https://github.com/$githubOrg/$githubProject")
+        }
+        project.addGithubPublishing(githubOrg, githubProject)
+    }
 
-    // Bintray publishing
-    var bintrayOrg: String? by project.extra
-    var bintrayUser: String? by project.extra
-    var bintrayApiKey: String? by project.extra
-    var bintrayRepo: String? by project.extra
+    /**
+     *  Space publishing
+     */
+    fun space(spaceRepo: String = "https://maven.pkg.jetbrains.space/mipt-npm/p/sci/maven") {
+        require(initializedFlag){"The publishing is not set up use 'setup' method to do so"}
+        project.addSpacePublishing(spaceRepo)
+    }
 
-    // Sonatype publising
-    var sonatypeUser: String? by project.extra
-    var sonatypePassword: String? by project.extra
+//    // Bintray publishing
+//    var bintrayOrg: String? by project.extra
+//    var bintrayUser: String? by project.extra
+//    var bintrayApiKey: String? by project.extra
+//    var bintrayRepo: String? by project.extra
+
+    /**
+     *  Sonatype publising
+     */
+    fun sonatype(){
+        require(initializedFlag){"The publishing is not set up use 'setup' method to do so"}
+        project.addSonatypePublishing()
+    }
 }
 
 
@@ -110,9 +133,9 @@ open class KScienceProjectPlugin : Plugin<Project> {
                 if (rootReadmeExtension.readmeTemplate.exists()) {
 
                     val modulesString = buildString {
-                        subprojects.forEach { subproject->
+                        subprojects.forEach { subproject ->
                             val name = subproject.name
-                            val path = subproject.path.replaceFirst(":","").replace(":","/")
+                            val path = subproject.path.replaceFirst(":", "").replace(":", "/")
                             val ext = subproject.extensions.findByType<KScienceReadmeExtension>()
                             appendln("<hr/>")
                             appendln("\n* ### [$name]($path)")