+++ src/main/kotlin/ru/mipt/npm/gradle/internal/common.kt
@@ -21,14 +21,11 @@ internal fun LanguageSettingsBuilder.applySettings(): Unit {
 
 internal fun RepositoryHandler.applyRepos(): Unit {
     mavenCentral()
+    maven("https://repo.kotlin.link")
     maven("https://dl.bintray.com/kotlin/kotlin-eap")
     maven("https://dl.bintray.com/kotlin/kotlin-dev")
     maven("https://kotlin.bintray.com/kotlinx")
     maven("https://kotlin.bintray.com/kotlin-js-wrappers/")
-    maven("https://dl.bintray.com/mipt-npm/kscience")
-    maven("https://dl.bintray.com/mipt-npm/dev")
-    maven("https://dl.bintray.com/mipt-npm/dataforge")
-    maven("https://repo.kotlin.link")
 }
 
 internal fun Copy.fromJsDependencies(configurationName: String) = project.afterEvaluate {